COVIDReport/requirements.txt
asgiref
cachetools
certifi
chardet
Django
google-api-core
googleapis-common-protos
google-auth
google-cloud-core
google-cloud-datastore
google-cloud-translate
grpcio
idna
isort
joblib
langdetect
mccabe
nltk
numpy
Package
pip
protobuf
pyasn1
pyasn1-modules
pytz
requests
rsa
scikit-learn
scipy
setuptools
six
sklearn
sqlparse
urllib3
wrapt
xlrd
xlwt
